requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
distro==1.9.0
Flask==2.3.2
flask-cors==6.0.0
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.79.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pydantic==2.11.4
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.42
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
